Casi esta

+++ Position.cpp
@@ -16,7 +16,7 @@ Position::Position()
 float Position::dist2(Position pos) 
 {
 
-	return sqrt(pow(this->x - pos.x, 2) + pow(this->y - pos.y, 2));
+	return sqrtf(pow(this->x - pos.x, 2) + pow(this->y - pos.y, 2));
 
 }
 
@@ -25,6 +25,6 @@ void Position::point2(Position pos)
 
 	float dx = pos.x - this->x;
 	float dy = pos.y - this->y;
-	this->direction = RAD2DEG(atan2f(dy, dx));
+	this->direction = (float)RAD2DEG(atan2f(dy, dx));
 
 }
